Move and fix imports for the new domain structure
